refactor: move qa into test section
